services/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wetpancake-backend"
version = "0.1.0"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "sqlalchemy>=2.0.0",
  "python-dotenv>=1.0.0",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]